ocfemia-lizelle-2023-hw7/requirements.txt
matplotlib
pytest
pyyaml
tqdm
scikit-learn
datasets
einops
scipy